fix(33-06): widen DebugHelp call-stack entries uint32->uintptr_t (A1-DBGHELP-RIP)
Resolve A1-DBGHELP-RIP by WIDENING (reviews fix #10a, preferred over
document-truncate): an x64 crash handler must not narrow the captured RIP.

- DebugHelp.h/.cpp: getCallStack output + lookupAddress address uint32 -> uintptr_t;
  the x64 RtlCaptureContext store now keeps the full 64-bit Rip (no DWORD() narrow)
- Cascade the wider entry through every caller buffer: CallStack[], CallStackCollector
  (storage + crc-of-bytes + heap copy), Fatal.cpp crash-handler buffer, RenderWorld
  DO_OBJECT_TRACKING callers[], MemoryManager owners[] + AllocatedBlock::m_owner/
  getOwner/setOwner (DO_TRACK block-owner storage)
- diagnostic %08X-with-cast<int> formats -> %p + void const* (no truncation)
- uintptr_t is 4 bytes on 32-bit == the prior uint32: Win32 ABI/behavior byte-unchanged
  (Win32 5-target relink clean, machine 14C); x64 relink clean (0 unresolved), 8664

A1-DBGHELP-WALK comment updated to point at the 33-06 controlled stack-walk gate.

+++ src/engine/shared/library/sharedDebug/src/shared/CallStack.h
@@ -11,6 +11,8 @@
 #ifndef INCLUDED_CallStack_H
 #define INCLUDED_CallStack_H
 
+#include <cstdint>   // PHASE-33 (A1-DBGHELP-RIP): uintptr_t for the call-stack entry type
+
 // ======================================================================
 
 class CallStack
@@ -40,7 +42,10 @@ private:
 
 private:
 
-	uint32 m_callStack[S_callStack];
+	// PHASE-33 (A1-DBGHELP-RIP): uintptr_t entries so the full 64-bit Rip survives
+	// on x64 (DebugHelp::getCallStack now writes uintptr_t). operator< memcmp uses
+	// sizeof(m_callStack), so it scales with the wider element automatically.
+	uintptr_t m_callStack[S_callStack];
 };
 
 // ======================================================================
